refactor: move find_pages_that_use_file into nextjs utils
